Main: warnings
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@52959 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Main/src-server/VRDEServerImpl.cpp
@@ -292,7 +292,7 @@ static int i_portParseNumber(uint16_t *pu16Port, const char *pszStart, const cha
      *       only digits and pszEnd points to the char after last
      *       digit.
      */
-    int cch = pszEnd - pszStart;
+    size_t cch = pszEnd - pszStart;
     if (cch > 0 && cch <= 5) /* Port is up to 5 decimal digits. */
     {
         unsigned uPort = 0;
